requirements.txt
python-telegram-bot==20.3
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0